# openapi.snapshot.json
{"components":{"schemas":{"ChunkRow":{"description":"Single row from the chunks table.","properties":{"char_count":{"title":"Char Count","type":"integer"},"chunk_id":{"title":"Chunk Id","type":"string"},"chunk_index":{"title":"Chunk Index","type":"integer"},"created_at":{"title":"Created At","type":"string"},"document_id":{"title":"Document Id","type":"string"},"id":{"title":"Id","type":"string"},"page_end":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Page End"},"page_start":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Page Start"},"section_path":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Section Path"},"text":{"title":"Text","type":"string"},"token_count":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Token Count"}},"required":["id","document_id","chunk_index","chunk_id","text","char_count","created_at"],"title":"ChunkRow","type":"object"},"DiagnosticsModel":{"description":"Deterministic rule-based inspection findings.\n\nAll counts exactly match the lengths of the returned row collections.\n``possible_issues`` is a stably sorted list of machine-friendly strings.","properties":{"chunk_count":{"minimum":0.0,"title":"Chunk Count","type":"integer"},"document_found":{"title":"Document Found","type":"boolean"},"embedding_count":{"minimum":0.0,"title":"Embedding Count","type":"integer"},"file_count":{"minimum":0.0,"title":"File Count","type":"integer"},"has_chunks":{"title":"Has Chunks","type":"boolean"},"has_embeddings":{"title":"Has Embeddings","type":"boolean"},"has_files":{"title":"Has Files","type":"boolean"},"has_summaries":{"title":"Has Summaries","type":"boolean"},"ingest_run_count":{"minimum":0.0,"title":"Ingest Run Count","type":"integer"},"possible_issues":{"items":{"type":"string"},"title":"Possible Issues","type":"array"},"summary_count":{"minimum":0.0,"title":"Summary Count","type":"integer"}},"required":["document_found","has_files","has_chunks","has_embeddings","has_summaries","ingest_run_count","chunk_count","embedding_count","summary_count","file_count","possible_issues"],"title":"DiagnosticsModel","type":"object"},"DocumentFileRow":{"description":"Single row from the document_files table.","properties":{"created_at":{"title":"Created At","type":"string"},"document_id":{"title":"Document Id","type":"string"},"file_ext":{"title":"File Ext","type":"string"},"id":{"title":"Id","type":"string"},"rel_path":{"title":"Rel Path","type":"string"},"role":{"title":"Role","type":"string"},"sha256":{"title":"Sha256","type":"string"},"size_bytes":{"title":"Size Bytes","type":"integer"}},"required":["id","document_id","role","rel_path","file_ext","sha256","size_bytes","created_at"],"title":"DocumentFileRow","type":"object"},"DocumentRow":{"description":"Single row from the documents table.","properties":{"author":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Author"},"content_sha256":{"title":"Content Sha256","type":"string"},"doc_type":{"title":"Doc Type","type":"string"},"document_id":{"title":"Document Id","type":"string"},"group_id":{"title":"Group Id","type":"string"},"ingest_version":{"title":"Ingest Version","type":"string"},"ingested_at":{"title":"Ingested At","type":"string"},"language":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Language"},"library_id":{"title":"Library Id","type":"string"},"mime_type":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Mime Type"},"original_filename":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Original Filename"},"page_count":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Page Count"},"security_level":{"title":"Security Level","type":"string"},"size_bytes":{"title":"Size Bytes","type":"integer"},"source_uri":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Source Uri"},"status":{"title":"Status","type":"string"},"title":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Title"}},"required":["document_id","library_id","group_id","security_level","doc_type","ingested_at","ingest_version","content_sha256","size_bytes","status"],"title":"DocumentRow","type":"object"},"EmbeddingRow":{"description":"Single row from the embeddings table.\n\nThe ``chunk_id`` field is the UUID primary key of the parent chunk\n(``chunks.id``), not the stable text ``chunk_id`` column.\n\nThe embedding vector is intentionally omitted — 2048 floats would dwarf\nall other diagnostic data and are not needed for pipeline inspection.","properties":{"chunk_id":{"title":"Chunk Id","type":"string"},"created_at":{"title":"Created At","type":"string"},"embedding_dim":{"title":"Embedding Dim","type":"integer"},"embedding_model":{"title":"Embedding Model","type":"string"},"id":{"title":"Id","type":"string"}},"required":["id","chunk_id","embedding_model","embedding_dim","created_at"],"title":"EmbeddingRow","type":"object"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"title":"Detail","type":"array"}},"title":"HTTPValidationError","type":"object"},"IngestRunRow":{"description":"Single row from the ingest_runs table.\n\n``document_id`` is nullable in the schema (run may not yet be associated\nwith a produced document).  ``metrics_json`` is stored as JSONB and\nreturned as a dict.","properties":{"document_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Document Id"},"ended_at":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Ended At"},"error_message":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Error Message"},"group_id":{"title":"Group Id","type":"string"},"id":{"title":"Id","type":"string"},"library_id":{"title":"Library Id","type":"string"},"metrics_json":{"additionalProperties":true,"title":"Metrics Json","type":"object"},"requested_by":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Requested By"},"security_level":{"title":"Security Level","type":"string"},"started_at":{"title":"Started At","type":"string"},"status":{"title":"Status","type":"string"}},"required":["id","library_id","group_id","security_level","status","started_at"],"title":"IngestRunRow","type":"object"},"InspectionResponse":{"description":"Canonical forensic inspection record for one document.\n\nWhen the document does not exist in TechVault, ``document`` is ``null``\nand all collection fields are empty arrays.  The response is still valid\nand deterministic — ``diagnostics.possible_issues`` will contain\n``\"document_missing\"``.\n\nOrdering rules (enforced by the service layer):\n- document_files: ascending by ``id``\n- chunks: ascending by ``chunk_index``, then ``id``\n- embeddings: ascending by ``chunk_id``, then ``id``\n- summaries: ascending by ``id``\n- ingest_runs: descending by ``started_at``, then ascending by ``id``","properties":{"chunks":{"items":{"$ref":"#/components/schemas/ChunkRow"},"title":"Chunks","type":"array"},"diagnostics":{"$ref":"#/components/schemas/DiagnosticsModel"},"document":{"anyOf":[{"$ref":"#/components/schemas/DocumentRow"},{"type":"null"}]},"document_files":{"items":{"$ref":"#/components/schemas/DocumentFileRow"},"title":"Document Files","type":"array"},"embeddings":{"items":{"$ref":"#/components/schemas/EmbeddingRow"},"title":"Embeddings","type":"array"},"ingest_runs":{"items":{"$ref":"#/components/schemas/IngestRunRow"},"title":"Ingest Runs","type":"array"},"query":{"$ref":"#/components/schemas/QueryInfo"},"summaries":{"items":{"$ref":"#/components/schemas/SummaryRow"},"title":"Summaries","type":"array"}},"required":["query","document_files","chunks","embeddings","summaries","ingest_runs","diagnostics"],"title":"InspectionResponse","type":"object"},"QueryInfo":{"description":"Echo of the lookup key used to produce this inspection record.","properties":{"document_id":{"title":"Document Id","type":"string"}},"required":["document_id"],"title":"QueryInfo","type":"object"},"SummaryRow":{"description":"Single row from the summaries table.","properties":{"content":{"title":"Content","type":"string"},"content_sha256":{"title":"Content Sha256","type":"string"},"created_at":{"title":"Created At","type":"string"},"document_id":{"title":"Document Id","type":"string"},"id":{"title":"Id","type":"string"},"model_used":{"title":"Model Used","type":"string"},"output_type":{"title":"Output Type","type":"string"},"provider":{"title":"Provider","type":"string"},"security_level":{"title":"Security Level","type":"string"},"summary_prompt_id":{"title":"Summary Prompt Id","type":"string"}},"required":["id","document_id","summary_prompt_id","output_type","model_used","provider","security_level","content","content_sha256","created_at"],"title":"SummaryRow","type":"object"},"ValidationError":{"properties":{"ctx":{"title":"Context","type":"object"},"input":{"title":"Input"},"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"title":"Location","type":"array"},"msg":{"title":"Message","type":"string"},"type":{"title":"Error Type","type":"string"}},"required":["loc","msg","type"],"title":"ValidationError","type":"object"}}},"info":{"description":"Deterministic forensic inspection of a single TechVault document record.","title":"document_record_inspector","version":"0.1.0"},"openapi":"3.1.0","paths":{"/health":{"get":{"operationId":"health_health_get","responses":{"200":{"content":{"application/json":{"schema":{"additionalProperties":{"type":"string"},"title":"Response Health Health Get","type":"object"}}},"description":"Successful Response"}},"summary":"Health","tags":["tools:document_record_inspector"]}},"/records/{document_id}":{"get":{"description":"Return the canonical inspection record for *document_id*.\n\nAlways returns HTTP 200.  When the document does not exist the response\nhas ``document: null`` and ``diagnostics.possible_issues`` contains\n``\"document_missing\"``.","operationId":"inspect_records__document_id__get","parameters":[{"in":"path","name":"document_id","required":true,"schema":{"title":"Document Id","type":"string"}}],"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/InspectionResponse"}}},"description":"Successful Response"},"422":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}},"description":"Validation Error"}},"summary":"Inspect","tags":["tools:document_record_inspector"]}}}}